+++ Packages/KeyKeyEngine/Tests/KeyKeyEngineTests/SmokeTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import KeyKeyEngine
-
-final class SmokeTests: XCTestCase {
-    func testVersionExists() {
-        XCTAssertEqual(KeyKeyEngine.version, "0.1.0")
-    }
-}